Refactor code to isolate SwiftSubtitles to our model
