+++ podcasts/TranscriptManager.swift
@@ -1,44 +1,29 @@
 import Foundation
-import SwiftSubtitles
 import PocketCastsDataModel
+
+enum TranscriptError: Error {
+    case notAvailable
+    case failedToLoad
+    case notSupported(format: String)
+    case failedToParse
+
+    var localizedDescription: String {
+        switch self {
+        case .notAvailable:
+            return "Transcript not available"
+        case .failedToLoad:
+            return "Transcript failed to load"
+        case .notSupported(let format):
+            return "Transcript format not supported: \(format)"
+        case .failedToParse:
+            return "Transcript failed to parse"
+        }
+    }
+}
 
 class TranscriptManager {
 
     typealias Transcript = Episode.Metadata.Transcript
-
-    enum TranscriptFormat: String {
-        case srt = "application/srt"
-        case vtt = "text/vtt"
-
-        var fileExtension: String {
-            switch self {
-            case .srt:
-                return "srt"
-            case .vtt:
-                return "vtt"
-            }
-        }
-    }
-
-    // Transcript formats we support in order of priority of use
-    static let supportedFormats: [TranscriptFormat] = [.srt, .vtt]
-
-    enum TranscriptError: Error {
-        case notAvailable
-        case failedToLoad
-        case notSupported(format: String)
-
-        var localizedDescription: String {
-            switch self {
-            case .notAvailable:
-                return "Transcript not available"
-            case .failedToLoad:
-                return "Transcript failed to load"
-            case .notSupported(let format):
-                return "Transcript format not supported: \(format)"
-            }
-        }
-    }
 
     let playbackManager: PlaybackManager
 
@@ -47,7 +32,7 @@ class TranscriptManager {
     }
 
     private func bestTranscript(from available: [Transcript]) -> Transcript? {
-        for format in Self.supportedFormats {
+        for format in TranscriptFormat.supportedFormats {
             if let transcript = available.first(where: { $0.type == format.rawValue}) {
                 return transcript
             }
@@ -74,9 +59,9 @@ class TranscriptManager {
             throw TranscriptError.failedToLoad
         }
 
-        let subtitles = try Subtitles(content: transcriptText, expectedExtension: transcriptFormat.fileExtension)
-
-        let model = TranscriptModel.makeModel(from: subtitles)
+        guard let model = TranscriptModel.makeModel(from: transcriptText, format: transcriptFormat) else {
+            throw TranscriptError.failedToParse
+        }
 
         return model
     }
@@ -84,7 +69,7 @@ class TranscriptManager {
 
 extension Episode.Metadata.Transcript {
 
-    var transcriptFormat: TranscriptManager.TranscriptFormat? {
-        return TranscriptManager.TranscriptFormat(rawValue: self.type)
+    var transcriptFormat: TranscriptFormat? {
+        return TranscriptFormat(rawValue: self.type)
     }
 }
